requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncpg]
python-jose[cryptography]
passlib[bcrypt]
pyshp
pyproj
typer
gunicorn
asyncpg